rename the parser script
